# internal/api/v1/openapi3/swagger.json
{
    "components": {
        "schemas": {
...
            "fxcore_internal_api_v1_dto.ApiResponse-any": {
                "properties": {
                    "code": {
                        "description": "业务码，见 internal/middleware/error.go 矩阵",
                        "type": "integer"
                    },
                    "data": {
                        "description": "业务数据"
                    },
                    "message": {
                        "description": "人类可读消息",
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "description": "unix 毫秒",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_LoginResponse": {
                "properties": {
                    "code": {
                        "description": "业务码，见 internal/middleware/error.go 矩阵",
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.LoginResponse"
                    },
                    "message": {
                        "description": "人类可读消息",
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "description": "unix 毫秒",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_RefreshResponse": {
                "properties": {
                    "code": {
                        "description": "业务码，见 internal/middleware/error.go 矩阵",
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.RefreshResponse"
                    },
                    "message": {
                        "description": "人类可读消息",
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "description": "unix 毫秒",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "fxcore_internal_api_v1_dto.ErrorResponse": {
                "properties": {
                    "code": {
                        "description": "业务错误码（1001-1500 矩阵）",
                        "type": "integer"
                    },
                    "data": {
                        "additionalProperties": {},
                        "description": "1001 时字段级错误映射，其余为 null",
                        "type": "object"
                    },
                    "message": {
                        "description": "可读错误信息",
                        "type": "string"
                    },
                    "request_id": {
                        "description": "链路追踪 ID",
                        "type": "string"
                    },
                    "timestamp": {
                        "description": "Unix 毫秒",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.LoginRequest": {
                "properties": {
                    "email": {
                        "maxLength": 255,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 128,
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ],
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.LoginResponse": {
                "description": "业务数据",
                "properties": {
                    "access_token": {
                        "type": "string"
                    },
                    "refresh_token": {
                        "type": "string"
                    },
                    "sign_secret": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.UserDTO"
                    }
                },
                "type": "object"
            },
...
            "fxcore_internal_api_v1_dto.RefreshRequest": {
                "properties": {
                    "refresh_token": {
                        "type": "string"
                    }
                },
                "required": [
                    "refresh_token"
                ],
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.ResetPasswordRequest": {
                "properties": {
                    "password": {
                        "maxLength": 128,
                        "minLength": 8,
                        "type": "string"
                    },
                    "token": {
                        "description": "邮件重置链接中的一次性令牌",
                        "maxLength": 128,
                        "type": "string"
                    }
                },
                "required": [
                    "token",
                    "password"
                ],
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.ResetPasswordResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_ForgotPasswordResponse": {
                "properties": {
                    "code": {
                        "description": "业务码，见 internal/middleware/error.go 矩阵",
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ForgotPasswordResponse"
                    },
                    "message": {
                        "description": "人类可读消息",
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "description": "unix 毫秒",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_ResetPasswordResponse": {
                "properties": {
                    "code": {
                        "description": "业务码，见 internal/middleware/error.go 矩阵",
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ResetPasswordResponse"
                    },
                    "message": {
                        "description": "人类可读消息",
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "description": "unix 毫秒",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.ForgotPasswordRequest": {
                "properties": {
                    "email": {
                        "format": "email",
                        "maxLength": 255,
                        "type": "string"
                    }
                },
                "required": [
                    "email"
                ],
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.ForgotPasswordResponse": {
                "properties": {
                    "dev_link": {
                        "description": "仅 SMTP 未配置的 dev 模式返回",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "fxcore_internal_api_v1_dto.RefreshResponse": {
                "description": "业务数据",
                "properties": {
                    "access_token": {
                        "type": "string"
                    },
                    "refresh_token": {
                        "type": "string"
                    },
                    "sign_secret": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "fxcore_internal_api_v1_dto.UserDTO": {
                "properties": {
                    "avatar": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "nickname": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "info": {
        "description": "AI 驱动的加密货币交易智能体平台 API（v1）。认证：access token 存 HttpOnly Cookie；写操作需 X-Signature/X-Timestamp/X-Nonce 签名头（HMAC-SHA256，±5min，nonce 防重放）。",
        "title": "FXcore API",
        "version": "1.0"
    },
...
    "paths": {
        "/auth/login": {
            "post": {
                "description": "校验账密，下发 access token（HttpOnly Cookie）与 refresh token / sign_secret",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.LoginRequest",
                                        "summary": "body",
                                        "description": "登录凭据"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "登录凭据",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_LoginResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 登录限流（5 次/分钟/IP）"
                    }
                },
                "summary": "登录",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "description": "携带 refresh_token 时服务端吊销；始终清除 access Cookie",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.RefreshRequest",
                                        "summary": "body",
                                        "description": "refresh token（可选，携带则吊销）"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "refresh token（可选，携带则吊销）"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-any"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "登出",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "description": "向注册邮箱发送密码重置链接；邮箱不存在时同样返回成功（防枚举）。SMTP 未配置时为 dev 模式：链接写日志并随响应 dev_link 返回",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ForgotPasswordRequest",
                                        "summary": "body",
                                        "description": "注册邮箱"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "注册邮箱",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_ForgotPasswordResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 限流（5 次/分钟/IP）"
                    }
                },
                "summary": "忘记密码",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/reset-password": {
            "post": {
                "description": "携带邮件链接中的一次性令牌设置新密码；成功后令牌立即作废",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ResetPasswordRequest",
                                        "summary": "body",
                                        "description": "重置令牌 + 新密码（≥8 位）"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "重置令牌 + 新密码（≥8 位）",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_ResetPasswordResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误 / 令牌无效或已过期"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 限流（5 次/分钟/IP）"
                    }
                },
                "summary": "重置密码",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "轮换 refresh token：旧 token 作废并返回新 token + 新 access token + sign_secret",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.RefreshRequest",
                                        "summary": "body",
                                        "description": "旧 refresh token"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "旧 refresh token",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_RefreshResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1101 refresh token 无效或已轮换"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 刷新限流"
                    }
                },
                "summary": "刷新令牌",
                "tags": [
                    "auth"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:8080/api/v1"
        }
    ]
}